requirements.txt
# Core Framework
fastapi
uvicorn
pydantic

# ML & AI Libraries
torch>=2.0.0
torchvision>=0.15.0
torch-geometric
tensorflow>=2.13.0
scikit-learn>=1.3.0
lifelines>=0.27.0
opencv-python>=4.8.0
mediapipe>=0.10.0
pillow>=10.0.0
shap>=0.42.0
joblib>=1.3.0

# Google AI
google-generativeai>=0.3.0

# Data Processing
numpy
pandas
scipy

# Database
firebase-admin
neo4j
motor
sqlalchemy
psycopg2-binary

# Utilities
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
aiohttp
requests

# Development & Visualization
pytest>=7.4.0
black
flake8
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
